+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # 1.0.1
 
 * First version of library
-* Trace Builder
+* Trace Builder as Builder pattern or as Singleton